+++ src/lex/mod.rs
@@ -1,3 +1,2 @@
-pub mod lexical_analyzer;
-pub mod lexical_analyzer_impl;
+pub mod lexical_analysis;
 pub mod lex_spec;