+++ src/token/token.rs
@@ -1,10 +1,10 @@
-pub struct Token<'a, TType> {
+pub struct Token<TType> {
     pub t_type: TType,
-    pub text: &'a str,
+    pub text: String,
 }
 
-impl<'a, TType> Token<'a, TType> {
-    pub fn new(t_type: TType, text: &'a str) -> Self {
+impl<TType> Token<TType> {
+    pub fn new(t_type: TType, text: String) -> Self {
         Token { t_type, text }
     }
 }
@@ -18,7 +18,7 @@ mod test {
         const TOKEN_TYPE: &str = "sample_type";
         const TOKEN_TEXT: &str = "sample_text";
 
-        let token: Token<&str> = Token::new(TOKEN_TYPE, TOKEN_TEXT);
+        let token: Token<&str> = Token::new(TOKEN_TYPE, String::from(TOKEN_TEXT));
 
         assert_eq!(token.t_type, TOKEN_TYPE);
         assert_eq!(token.text, TOKEN_TEXT);